Refactor imports of interfaces, entities and util functions
