+++ nodes/Stellar/actions/claimable/claimClaimableBalance/description.ts
@@ -1,4 +1,4 @@
-import { ClaimableBalanceProperties } from '../../entities/IStellarNode';
+import { ClaimableBalanceProperties } from '../../entities/StellarNode';
 
 export const claimClaimableBalanceDescription: ClaimableBalanceProperties = [
 	{
