+++ nodes/Stellar/actions/clawback/clawback/execute.ts
@@ -1,7 +1,9 @@
 import { IExecuteFunctions } from 'n8n-workflow';
-import { Operation } from 'stellar-sdk';
-import { buildAsset, convertAmountToBigNumber } from '../../../transport';
-import IAsset from '../../entities/IAsset';
+import { Asset, Operation } from 'stellar-sdk';
+import { convertAmountToBigNumber } from '../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
+import IAsset from '../../../../../common/interfaces/stellarBlockchain/IAsset';
+import { NetworkEnum } from '../../../../../common/enum/stellarBlockchain/networkEnum';
+import { buildAsset } from '../../../../../common/utils/stellarBlockchain/buildAsset';
 
 export async function clawback(this: IExecuteFunctions) {
 	try {
@@ -10,7 +12,7 @@ export async function clawback(this: IExecuteFunctions) {
 		const amountToBurn = this.getNodeParameter('amount', 0) as number;
 
 		const amount = convertAmountToBigNumber(amountToBurn);
-		const asset = buildAsset(assetToBurn);
+		const asset = buildAsset(assetToBurn, NetworkEnum.STELLAR) as Asset;
 		const clawbackOperation = Operation.clawback({ asset, amount, from }).toXDR('base64');
 
 		return { operation: clawbackOperation };
