+++ nodes/Stellar/actions/liquidityPool/liquidityPoolDeposit/helpers/helpers.ts
@@ -1,6 +1,6 @@
-import { convertAmountToBigNumber } from '../../../../transport';
-import ILiquidityPoolPrice from '../../../entities/ILiquidityPoolPrice';
-import IPriceFraction from '../../../entities/IPriceFraction';
+import { convertAmountToBigNumber } from '../../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
+import ILiquidityPoolPrice from '../../../../../../common/interfaces/stellarBlockchain/ILiquidityPoolPrice';
+import IPriceFraction from '../../../../../../common/interfaces/stellarBlockchain/IPriceFraction';
 import NoPriceSelectedError from '../errors/NoPriceSelectedError';
 
 export default function getPrice(price: ILiquidityPoolPrice['values']): string | IPriceFraction {
