+++ nodes/Stellar/actions/liquidityPool/liquidityPoolWithdraw/execute.ts
@@ -1,6 +1,6 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import { Operation } from 'stellar-sdk';
-import { convertAmountToBigNumber } from '../../../transport';
+import { convertAmountToBigNumber } from '../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
 
 export async function liquidityPoolWithdraw(this: IExecuteFunctions) {
 	try {
