+++ nodes/Stellar/actions/offers/manageBuyOffer/execute.ts
@@ -1,7 +1,9 @@
 import { IExecuteFunctions } from 'n8n-workflow';
-import { Operation } from 'stellar-sdk';
-import { buildAsset, convertAmountToBigNumber } from '../../../transport';
-import IAsset from '../../entities/IAsset';
+import { Operation, Asset } from 'stellar-sdk';
+import { convertAmountToBigNumber } from '../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
+import IAsset from '../../../../../common/interfaces/stellarBlockchain/IAsset';
+import { buildAsset } from '../../../../../common/utils/stellarBlockchain/buildAsset';
+import { NetworkEnum } from '../../../../../common/enum/stellarBlockchain/networkEnum';
 
 export async function manageBuyOffer(this: IExecuteFunctions) {
 	try {
@@ -11,8 +13,8 @@ export async function manageBuyOffer(this: IExecuteFunctions) {
 		const sellingPrice = this.getNodeParameter('price', 0) as number;
 		const offerId = this.getNodeParameter('offerId', 0) as string;
 
-		const selling = buildAsset(sellingAsset);
-		const buying = buildAsset(buyingAsset);
+		const selling = buildAsset(sellingAsset, NetworkEnum.STELLAR) as Asset;
+		const buying = buildAsset(buyingAsset, NetworkEnum.STELLAR) as Asset;
 		const buyAmount = convertAmountToBigNumber(sellingAmount);
 		const price = convertAmountToBigNumber(sellingPrice);
