+++ nodes/Stellar/actions/payments/pathPaymentStrictReceive/execute.ts
@@ -1,8 +1,10 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import { Asset, Operation } from 'stellar-sdk';
-import { buildAsset, convertAmountToBigNumber } from '../../../transport';
-import IAssetsPath from '../../entities/IAssetsPath';
-import IAsset from '../../entities/IAsset';
+import { convertAmountToBigNumber } from '../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
+import IAssetsPath from '../../../../../common/interfaces/stellarBlockchain/IAssetsPath';
+import IAsset from '../../../../../common/interfaces/stellarBlockchain/IAsset';
+import { buildAsset } from '../../../../../common/utils/stellarBlockchain/buildAsset';
+import { NetworkEnum } from '../../../../../common/enum/stellarBlockchain/networkEnum';
 
 export async function pathPaymentStrictReceive(this: IExecuteFunctions) {
 	try {
@@ -17,14 +19,14 @@ export async function pathPaymentStrictReceive(this: IExecuteFunctions) {
 			[],
 		) as IAssetsPath;
 
-		const sendAsset = buildAsset(sendingAsset);
+		const sendAsset = buildAsset(sendingAsset, NetworkEnum.STELLAR) as Asset;
 		const sendMax = convertAmountToBigNumber(maxSendingAmount);
-		const destAsset = buildAsset(destinationAsset);
+		const destAsset = buildAsset(destinationAsset, NetworkEnum.STELLAR) as Asset;
 		const destAmount = convertAmountToBigNumber(destinationAmount);
 		let path: Asset[] = [];
 
 		intermediateAssets.forEach((asset) => {
-			const intermediateAsset = buildAsset(asset);
+			const intermediateAsset = buildAsset(asset, NetworkEnum.STELLAR) as Asset;
 			path.push(intermediateAsset);
 		});
