+++ nodes/Stellar/actions/server/findPayments/description.ts
@@ -1,4 +1,4 @@
-import { ServerProperties } from '../../entities/IStellarNode';
+import { ServerProperties } from '../../entities/StellarNode';
 
 export const findPaymentsDescription: ServerProperties = [
 	{
