+++ nodes/Stellar/actions/settings/bumpSequence/description.ts
@@ -1,4 +1,4 @@
-import { SettingsProperties } from '../../entities/IStellarNode';
+import { SettingsProperties } from '../../entities/StellarNode';
 
 export const bumpSequenceDescription: SettingsProperties = [
 	{
