+++ nodes/Stellar/actions/sponsorship/revokeSponsorship/revokeTrustlineSponsorship/description.ts
@@ -1,4 +1,4 @@
-import { SponsorshipProperties } from '../../../entities/IStellarNode';
+import { SponsorshipProperties } from '../../../entities/StellarNode';
 
 export const revokeTrustlineSponsorshipDescription: SponsorshipProperties = [
 	{
