+++ nodes/Stellar/actions/trust/changeTrust/execute.ts
@@ -1,7 +1,9 @@
 import { IExecuteFunctions } from 'n8n-workflow';
 import { Asset, LiquidityPoolAsset, Operation } from 'stellar-sdk';
-import { buildAsset, convertAmountToBigNumber } from '../../../transport';
-import IAsset from '../../entities/IAsset';
+import { convertAmountToBigNumber } from '../../../../../common/utils/stellarBlockchain/convertAmountToBigNumber';
+import IAsset from '../../../../../common/interfaces/stellarBlockchain/IAsset';
+import { NetworkEnum } from '../../../../../common/enum/stellarBlockchain/networkEnum';
+import { buildAsset } from '../../../../../common/utils/stellarBlockchain/buildAsset';
 
 export async function changeTrust(this: IExecuteFunctions) {
 	try {
@@ -20,8 +22,8 @@ export async function changeTrust(this: IExecuteFunctions) {
 				const { values: assetBValues } = this.getNodeParameter('assetB', 0) as IAsset;
 				const fee = this.getNodeParameter('fee', 0) as number;
 
-				const assetA = buildAsset(assetAValues);
-				const assetB = buildAsset(assetBValues);
+				const assetA = buildAsset(assetAValues, NetworkEnum.STELLAR) as Asset;
+				const assetB = buildAsset(assetBValues, NetworkEnum.STELLAR) as Asset;
 
 				asset = new LiquidityPoolAsset(assetA, assetB, fee);
 				break;
